pyproject.toml
[tool.poetry]
name = "owl2yarrrml"
version = "1.2"
description = "A Python library for translating OWL ontologies into YARRRML mappings"
authors = ["David Chaves <david.chaves@usc.es>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/citius/owl2yarrrml"
repository = "https://github.com/citius/owl2yarrrml"
keywords = ["owl", "yarrrml", "rdf", "mapping", "ontology"]

[tool.poetry.dependencies]
python = ">3.8.1,<4.0.0"
PyYAML = "^6.0"
owlready2 = "^0.39.0"
rdflib="^7.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.3.2"]
build-backend = "poetry.core.masonry.api"
